move cobra files to their own subfolder
