+++ ws/ixcobra/IXCobraConnection.cpp
